Add repeating test ChannelSinkTaskTest to ensure records are received in order
Adds a repeating test to assert that records are received with correct size and order. Time intervals for putting sink records and connection state changes were chosen randomly.  That is to catch any possible race condition and missing records. I also chose to make this a repeating test so that in each test run, different random time intervals are chosen. That increases the confidence in functionality